# Sistema_Gestión_Donante/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Donantes - API",
        "description": "API para el módulo de donantes del sistema de adopciones de animales",
        "contact": {
            "email": "admin@sistema-adopciones.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/verificaciones-donantes": {
            "get": {
                "tags": [
                    "Verificaciones Donantes"
                ],
                "summary": "Obtener todas las verificaciones de donantes",
                "operationId": "5fa53728a9c113d0bf9bd1b3524e0d1f",
                "responses": {
                    "200": {
                        "description": "Lista de verificaciones de donantes obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Verificaciones Donantes"
                ],
                "summary": "Crear nueva verificación de donante",
                "operationId": "c5eeedcbb3bb2cef148448043d470273",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "donante_id",
                                    "fecha_verificacion",
                                    "resultado"
                                ],
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_verificacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-08"
                                    },
                                    "resultado": {
                                        "type": "string",
                                        "example": "Aprobado"
                                    },
                                    "observaciones": {
                                        "type": "string",
                                        "example": "Documentos en orden, domicilio verificado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Verificación de donante creada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/verificaciones-donantes/{id}": {
            "get": {
                "tags": [
                    "Verificaciones Donantes"
                ],
                "summary": "Obtener verificación de donante por ID",
                "operationId": "799e87831fd2b41b240df37657cc0b39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la verificación de donante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verificación de donante obtenida exitosamente"
                    },
                    "404": {
                        "description": "Verificación de donante no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Verificaciones Donantes"
                ],
                "summary": "Actualizar verificación de donante existente",
                "operationId": "08b5779644b4fed9144fe1fb14dcc00f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la verificación de donante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_verificacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-08"
                                    },
                                    "resultado": {
                                        "type": "string",
                                        "example": "Observación"
                                    },
                                    "observaciones": {
                                        "type": "string",
                                        "example": "Requiere documentación adicional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verificación de donante actualizada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Verificación de donante no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Verificaciones Donantes"
                ],
                "summary": "Eliminar verificación de donante",
                "operationId": "aa26d22b2e6843929ae91c8c47acb271",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la verificación de donante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verificación de donante eliminada exitosamente"
                    },
                    "404": {
                        "description": "Verificación de donante no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/verificaciones-donantes/donante/{donanteId}": {
            "get": {
                "tags": [
                    "Verificaciones Donantes"
                ],
                "summary": "Obtener verificaciones por donante",
                "operationId": "bd668ff3301c2655f76c3f80da52b9fd",
                "parameters": [
                    {
                        "name": "donanteId",
                        "in": "path",
                        "description": "ID del donante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de verificaciones por donante obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/verificaciones-donantes/resultado/{resultado}": {
            "get": {
                "tags": [
                    "Verificaciones Donantes"
                ],
                "summary": "Obtener verificaciones por resultado",
                "operationId": "0013c63753313db98d8b9f9c9d994849",
                "parameters": [
                    {
                        "name": "resultado",
                        "in": "path",
                        "description": "Resultado de la verificación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de verificaciones por resultado obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        }
    },
...
}